pyproject.toml
[tool.poetry]
name = "agents-crewai-example"
version = "0.1"
description = "Example of using crewAI agents with open source LLM"
authors = ["artem"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
pydantic = "^2.4.2"
langchain = "^0.1.10"
openai = "^1.13.3"
regex = "^2023.12.25"
crewai-tools = "^0.2.6"
python-dotenv = "^1.0.0"
crewai = {git = "git@github.com:madmag77/crewAI.git"}